+++ packages/plugins/AddressSearch/src/store/getters.ts
@@ -62,7 +62,7 @@ const getters: PolarGetterTree<AddressSearchState, AddressSearchGetters> = {
       ...(rootGetters.configuration?.addressSearch || {}),
     }
   },
-  afterResultComponent(_, { addressSearchConfiguration }): Vue | null {
+  afterResultComponent(_, { addressSearchConfiguration }) {
     return addressSearchConfiguration.afterResultComponent || null
   },
   minLength(_, { addressSearchConfiguration }) {