+++ packages/types/custom/core.ts
@@ -127,6 +127,7 @@ export interface AddressSearchConfiguration extends PluginOptions {
   addLoading?: string
   // optional loading action name to end loading
   removeLoading?: string
+  afterResultComponent?: Vue
 }
 
 export interface Attribution {